feat: Added Lazygit for Neovim
Added Lazygit for git client. I already use it on the terminal, so
I wanted to include it in Neovim. This means that Neogit was removed.
I've also updated my aliases to start telescope in the current directory
for neovim.

breaking: removed Neogit
breaking: changed 'n' alias to start telescope